+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.0
+
+### Added
+
+- Added `ConsistentRandom.testing` method to allow for deterministic testing of random values.
+- Added `ConsistentRandom#name` method to return the name used for seeding the random value.
+
 ## 2.1.1
 
 ### Fixed